+++ babel.config.js
@@ -3,5 +3,6 @@ module.exports = {
         ['@babel/preset-env', {targets: {node: 'current'}}],
         ['@babel/preset-flow', {targets: {node: 'current'}}],
         '@babel/preset-typescript',
+        ['@babel/preset-react', {runtime: 'automatic'}],
     ],
   };